+++ src/lib/database.js
@@ -465,68 +465,115 @@ async function search(parameters, page, limit = 10) {
   return response
 }
 
-async function aggregate(parameters) {
+const ALL_AGGREGATIONS = {
+  total_count: { value_count: { field: 'id' } },
+  collection_frequency: { terms: { field: 'collection', size: 100 } },
+  platform_frequency: { terms: { field: 'properties.platform', size: 100 } },
+  cloud_cover_frequency: {
+    range: {
+      field: 'properties.eo:cloud_cover',
+      ranges: [
+        { to: 5 },
+        { from: 5, to: 15 },
+        { from: 15, to: 40 },
+        { from: 40 },
+      ],
+    }
+  },
+  datetime_frequency: {
+    date_histogram: {
+      field: 'properties.datetime',
+      calendar_interval: 'month',
+    }
+  },
+  datetime_min: { min: { field: 'properties.datetime' } },
+  datetime_max: { max: { field: 'properties.datetime' } },
+  grid_code_frequency: {
+    terms: {
+      field: 'properties.grid:code',
+      size: 2000,
+      missing: 'none',
+    }
+  },
+  grid_code_landsat_frequency: {
+    terms: {
+      field: 'properties.landsat:wrs_type',
+      size: 2000,
+      missing: 'none',
+      script: {
+        lang: 'painless',
+        source: "return 'WRS' + _value + '-' + "
+          + "doc['properties.landsat:wrs_path'].value + "
+          + "doc['properties.landsat:wrs_row'].value"
+      }
+    }
+  },
+  sun_elevation_frequency: {
+    histogram: {
+      field: 'properties.view:sun_elevation',
+      interval: 5
+    }
+  },
+  sun_azimuth_frequency: {
+    histogram: {
+      field: 'properties.view:sun_azimuth',
+      interval: 5
+    }
+  },
+  off_nadir_frequency: {
+    histogram: {
+      field: 'properties.view:off_nadir',
+      interval: 5
+    }
+  }
+}
+
+/**
+ * @param {Array} aggregations
+ * @param {Object} parameters
+ * @param {number} geohashPrecision
+ * @param {number} geohexPrecision
+ * @param {number} geotilePrecision
+ * @returns {Promise<Object>}
+ */
+async function aggregate(
+  aggregations, parameters,
+  geohashPrecision, geohexPrecision, geotilePrecision
+) {
   const searchParams = await constructSearchParams(parameters)
   searchParams.body.size = 0
-  searchParams.body.aggs = {
-    total_count: { value_count: { field: 'id' } },
-    collection_frequency: { terms: { field: 'collection', size: 100 } },
-    platform_frequency: { terms: { field: 'properties.platform', size: 100 } },
-    cloud_cover_frequency: {
-      range: {
-        field: 'properties.eo:cloud_cover',
-        ranges: [
-          { to: 5 },
-          { from: 5, to: 15 },
-          { from: 15, to: 40 },
-          { from: 40 },
-        ],
+
+  logger.debug('Aggregations : %j', aggregations)
+  // include all aggregations specified
+  // this will ignore aggregations with the wrong names
+  searchParams.body.aggs = Object.keys(ALL_AGGREGATIONS).reduce((o, k) => {
+    if (aggregations.includes(k)) o[k] = ALL_AGGREGATIONS[k]
+    return o
+  }, {})
+
+  if (aggregations.includes('grid_geohex_frequency')) {
+    searchParams.body.aggs.grid_geohex_frequency = {
+      geohex_grid: {
+        field: 'properties.proj:centroid',
+        precision: geohexPrecision
       }
-    },
-    datetime_frequency: {
-      date_histogram: {
-        field: 'properties.datetime',
-        calendar_interval: 'month',
+    }
+  }
+
+  if (aggregations.includes('grid_geohash_frequency')) {
+    searchParams.body.aggs.grid_geohash_frequency = {
+      geohash_grid: {
+        field: 'properties.proj:centroid',
+        precision: geohashPrecision
       }
-    },
-    datetime_min: { min: { field: 'properties.datetime' } },
-    datetime_max: { max: { field: 'properties.datetime' } },
-    grid_code_frequency: {
-      terms: {
-        field: 'properties.grid:code',
-        size: 1000,
-        missing: 'none',
-      }
-    },
-    grid_code_landsat_frequency: {
-      terms: {
-        field: 'properties.landsat:wrs_type',
-        size: 1000,
-        missing: 'none',
-        script: {
-          lang: 'painless',
-          source: "return 'WRS' + _value + '-' + "
-            + "doc['properties.landsat:wrs_path'].value + "
-            + "doc['properties.landsat:wrs_row'].value"
-        }
-      }
-    },
-    sun_elevation_frequency: {
-      histogram: {
-        field: 'properties.view:sun_elevation',
-        interval: 5
-      }
-    },
-    sun_azimuth_frequency: {
-      histogram: {
-        field: 'properties.view:sun_azimuth',
-        interval: 5
-      }
-    },
-    off_nadir_frequency: {
-      histogram: {
-        field: 'properties.view:off_nadir',
-        interval: 5
+    }
+  }
+
+  if (aggregations.includes('grid_geotile_frequency')) {
+    searchParams.body.aggs.grid_geotile_frequency = {
+      geotile_grid: {
+        field: 'properties.proj:centroid',
+        precision: geotilePrecision
       }
     }
   }