+++ CHANGELOG.md
@@ -16,11 +16,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 
 - Child link relations removed from root (landing page).
+- POST /aggregate endpoints were removed, as they didn't work correctly.
 
 ## Added
 
 - Added support for `/aggregations`, `/collections/{collectionId}/aggregations`, and
   `/collections/{collectionId}/aggregate` endpoints.
+- Added grid_geohex_frequency, grid_geohash_frequency, and grid_geotile_frequency aggregations
 
 ## Fixed